checkers: Move base Checker class into this module
