+++ src/checkers/gnomechecker.py
@@ -9,7 +9,7 @@ from ..lib.errors import CheckerQueryError
 from ..lib.externaldata import ExternalBase, ExternalFile
 from ..lib.checksums import MultiDigest
 from ..lib.utils import filter_versions
-from ..lib.checkers import Checker
+from . import Checker
 
 log = logging.getLogger(__name__)
