+++ src/checkers/jetbrainschecker.py
@@ -3,7 +3,7 @@ import logging
 
 from ..lib.externaldata import ExternalBase, ExternalFile
 from ..lib.checksums import MultiDigest
-from ..lib.checkers import Checker
+from . import Checker
 
 log = logging.getLogger(__name__)
