+++ src/checkers/rustchecker.py
@@ -6,7 +6,7 @@ import toml
 
 from ..lib.externaldata import ExternalBase, ExternalFile
 from ..lib.checksums import MultiDigest
-from ..lib.checkers import Checker
+from . import Checker
 
 log = logging.getLogger(__name__)
