+++ src/checkers/snapcraftchecker.py
@@ -4,7 +4,7 @@ import hashlib
 
 from ..lib.externaldata import ExternalBase, ExternalFile
 from ..lib.checksums import MultiHash
-from ..lib.checkers import Checker
+from . import Checker
 
 log = logging.getLogger(__name__)
